// src/dto/user.dto.ts
export class UserDto {
  firstName: string;
  lastName: string;
  email: string;
  birthDate: string;
}

export class UserRequest {
  firstName: string;
  lastName: string;
  email: string;
  birthDate: string;
}

export class CreateUserDto {
  firstName: string;
  lastName: string;
  email: string;
  birthDate: string;
}